convert to gui

+++ mounting4Paul/mount_Digital_Design.sh
@@ -5,9 +5,9 @@
 ## Author:         Roman Willi
 ##
 ## Creation Date:  01.04.2025
-## Script Name:    unmount_Digital_Design.sh
+## Script Name:    mount_Digital_Design.sh
 ## Project Name:   
-## Description:    Unmountet Unterrichtsserver
+## Description:    Mountet Unterrichtsserver in user home
 ##
 ## Revision:
 ## Revision        01 - File Created
@@ -18,10 +18,44 @@ link_path=~/Digital_Design
 mount_path="smb://sifs09.ost.ch/bsc.et/unterricht/module_BSc/Digital_Design"
 mount_point="$XDG_RUNTIME_DIR/gvfs/smb-share:server=sifs09.ost.ch,share=bsc.et/unterricht/module_BSc/Digital_Design"
 
+export DBUS_SESSION_BUS_ADDRESS="unix:path=/run/user/$UID/bus"
+
 echo "Unmounting $mount_path..."
 gio mount -u "$mount_path"
+
 echo "Removing the link $link_path..."
-unlink "$link_path"
+[ -L "$link_path" ] && unlink "$link_path"
+
+echo "Mounting: $mount_path"
+gio mount "$mount_path"
+
+# Warte, bis Mountpoint wirklich da ist (max. 10 Sekunden)
+echo -n "Warte auf Mountpoint..."
+for i in {1..10}; do
+    if [ -d "$mount_point" ]; then
+        echo " gefunden."
+        break
+    fi
+    echo -n "."
+    sleep 1
+done
+
+if [ ! -d "$mount_point" ]; then
+    echo "❌ Fehler: Mountpoint nicht gefunden!"
+    exit 1
+fi
+
+echo "Creating symlink: $link_path"
+ln -s "$mount_point" "$link_path"
+
+if [ -e "$link_path" ]; then
+	echo "✅ The link $link_path is working"
+	code&
+else
+	echo "❌ The link $link_path is not working"
+	exit 1
+fi
+
 
 # Manual Use
 # ----------
@@ -33,3 +67,6 @@ unlink "$link_path"
 # UNMOUNT, UNLINK
 #gio mount -u smb://sifs09.ost.ch/bsc.et/unterricht/module_BSc/Digital_Design
 #unlink ~/Digital_Design
+
+# Auführbar machen:
+# chmod +x meinscript.sh
